adds happy and sad paths for equipment set create and update

+++ app/controllers/equipment_sets_controller.rb
@@ -7,23 +7,27 @@ class EquipmentSetsController < ApplicationController
 
   def create
     #Create actual equipment set
-    set = EquipmentSet.create(
+    set = EquipmentSet.new(
       name: params[:name],
       user_id: current_user.id,
       oldschooljs_monster_id: params[:oldschooljs_monster_id]
     )
-
-    #Create individual items to attach to equipment set
-    items = params[:items]
-    items.each do |item|
-      p item[:oldschooljs_item_id]
-      EquipmentSetItem.create(
-        oldschooljs_item_id: item[:oldschooljs_item_id],
-        slot: item[:slot],
-        equipment_set_id: set.id
-      )
+    if set.save
+      #Create individual items to attach to equipment set
+      items = params[:items]
+      items.each do |item|
+        p item[:oldschooljs_item_id]
+        EquipmentSetItem.create(
+          oldschooljs_item_id: item[:oldschooljs_item_id],
+          slot: item[:slot],
+          equipment_set_id: set.id
+        )
+      end
+      render json: set
+    else
+      render json: {errors: set.errors.full_messages}, status: :unprocessable_entity
     end
-    render json: set
+    
   end
 
   def show
@@ -35,20 +39,23 @@ class EquipmentSetsController < ApplicationController
     set = EquipmentSet.find(params[:equipment_set_id])
     set.name = params[:name] || set.name
     set.oldschooljs_monster_id = params[:oldschooljs_monster_id] || set.oldschooljs_monster_id
-    set.save
-
+    
+    if set.save
     #Deletes old items and adds new items
-    EquipmentSetItem.where(equipment_set_id: set.id).destroy_all
-    items = params[:items]
-    items.each do |item|
-      EquipmentSetItem.create(
-        oldschooljs_item_id: item[:oldschooljs_item_id],
-        slot: item[:slot],
-        equipment_set_id: set.id
-      )
-    end
+      EquipmentSetItem.where(equipment_set_id: set.id).destroy_all
+      items = params[:items]
+      items.each do |item|
+        EquipmentSetItem.create(
+          oldschooljs_item_id: item[:oldschooljs_item_id],
+          slot: item[:slot],
+          equipment_set_id: set.id
+        )
+      end
+      render json: set
+    else
+      render json: {errors: set.errors.full_messages}, status: :unprocessable_entity
 
-    render json: set
+    end
 
   end
